requirements.txt
Django==2.0.5
dj-database-url==0.4.2
djangorestframework==3.8.2
